+++ crates/ruma-common/src/serde.rs
@@ -12,6 +12,7 @@ use serde::{
     Deserialize, Deserializer,
 };
 use serde_json::{value::RawValue as RawJsonValue, Value as JsonValue};
+use tracing::debug;
 
 pub mod base64;
 mod buf;
@@ -76,6 +77,20 @@ where
     E: de::Error,
 {
     serde_json::from_str(val.get()).map_err(E::custom)
+}
+
+/// Helper function for returning a default value if deserialization of the type fails.
+///
+/// Used as `#[serde(deserialize_with = "default_on_error")]`.
+pub fn default_on_error<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+    T: Deserialize<'de> + Default,
+{
+    Ok(T::deserialize(deserializer).unwrap_or_else(|error| {
+        debug!("deserialization error, using default value: {error}");
+        T::default()
+    }))
 }
 
 /// Helper function for ignoring invalid items in a `Vec`, instead letting them cause the entire